requirements.txt
Flask
instaloader
openai-whisper
werkzeug